+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * adjusted default values to run out of the box
     * tls is optional now, run http only if not specified
     * storage is optional now, run in emptyDirs if not specified
+* print loaded books on startup
 
 ## 1.4.0 (2021-11-07)